Add a parse function to the library
This commit adjusts the specification to allow for the
inclusion of a version number in the type identifier.

It also adds a utility function to the library which
takes a string and parses it as a type identifier to
separate the different parts. This function is
exposed as `type.parse`.

Furthermore, because this change asked for a more
extensive API documentation, the commit adds validation
of the documentation by doctest.

Closes #4
Closes #8 (supersedes)